Restructured backend routing, added a route file and controllers, fileUpload WIP

+++ db/migrations/20180726181653_create_listings_table.js
@@ -7,8 +7,8 @@ exports.up = function(knex, Promise) {
     table.float('price');
     table.integer('landlords_id');
     table.foreign('landlords_id').references('landlords.id')
-    table.float('lng');
-    table.float('lat');
+    table.double('lng');
+    table.double('lat');
     table.integer('neighbourhoods_id');
     table.foreign('neighbourhoods_id').references('neighbourhoods.id');
   });
